backend/pyproject.toml
# =============================================================================
# Theatre Backend - Project Configuration
# =============================================================================

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "theatre-backend"
version = "0.1.0"
description = "Theatre Management System - Backend API"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Proprietary"}
authors = [
    {name = "Theatre Team"}
]

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    
    # Validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "email-validator>=2.1.0",
    
    # Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<4.1.0",  # Фиксируем версию для совместимости с passlib
    "python-multipart>=0.0.9",
    
    # Redis
    "redis>=5.0.0",
    
    # Utilities
    "aiofiles>=23.2.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    
    # Linting & Formatting
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    
    # Development utilities
    "ipython>=8.20.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

# =============================================================================
# Ruff Configuration
# =============================================================================
[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "alembic/versions",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (FastAPI Depends)
    "B904",   # raise without from (иногда нужно)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

# =============================================================================
# MyPy Configuration
# =============================================================================
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "redis.*",
    "jose.*",
    "passlib.*",
]
ignore_missing_imports = true

# =============================================================================
# Pytest Configuration
# =============================================================================
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/alembic/*"]
